// src/Utils/Time/types.ts
export type WeekDaysList =
	| "sunday"
	| "monday"
	| "tuesday"
	| "wednesday"
	| "thursday"
	| "friday"
	| "saturday";

export enum WeekDays {
	"sunday",
	"monday",
	"tuesday",
	"wednesday",
	"thursday",
	"friday",
	"saturday",
}

export type IGetTimeFormmated =
	| "DMY-HYPEN"
	| "DMY-SLASH"
	| "MY-HYPEN"
	| "MY-SLASH"
	| "YMD-HYPEN"
	| "YMD-SLASH";